Custom channels configuration
